refactor(testes): uma classe por arquivo e um so lugar para o banco em memoria
PersistenciaTeste.cs tinha 354 linhas e tres classes de teste sem relacao entre
si — o gatilho de tamanho do CLAUDE.md. `ResolvedorSobreBancoTeste` e
`FichaNoBancoTeste` saem para arquivos proprios; o que sobra volta a falar de
uma coisa so, com 154 linhas.

O setup do SQLite em memoria estava copiado em QUATRO classes: mesma conexao
aberta, mesmas opcoes, mesmo EnsureCreated, mesmo Dispose. A terceira aparicao
ja era o gatilho de DRY, e aqui a duplicacao nao e acidental — se o setup
mudar (outro provider, outra versao, um EnsureDeleted no meio), muda para todas
ao mesmo tempo. Virou `BancoEmMemoria`, classe base com `Opcoes` e
`NovoContexto()`.

`NovoContexto()` carrega o motivo de existir: contexto novo aproxima o teste de
um processo novo, porque o ChangeTracker nasce vazio e o que volta veio do
banco, nao da memoria de quem gravou. Era isso que as tres copias faziam sem
dizer.

Nenhum assert mudou: os testes sao os mesmos, so mudaram de arquivo e passaram
a herdar o setup.

Closes #111
