+++ testes/Copiloto.Testes/PersistenciaTeste.cs
@@ -3,9 +3,7 @@ using Copiloto.Api.Persistencia;
 using Copiloto.Dominio.Conversas;
 using Copiloto.Dominio.Ia;
 using Copiloto.Dominio.Vendas;
-using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
-using Fichas = Copiloto.Dominio.Fichas;
 
 namespace Copiloto.Testes;
 
@@ -17,30 +15,11 @@ namespace Copiloto.Testes;
 /// no primeiro clone. O que se prova aqui e o MAPEAMENTO — que o modelo fecha,
 /// que as chaves e os indices existem e que o dominio sobrevive a ida e volta.
 /// </summary>
-public class PersistenciaTeste : IDisposable
+public class PersistenciaTeste : BancoEmMemoria
 {
     private static readonly DateTimeOffset Agora = new(2026, 9, 1, 12, 0, 0, TimeSpan.Zero);
 
-    private readonly SqliteConnection _conexao;
-    private readonly DbContextOptions<CopilotoDbContext> _opcoes;
-
-    public PersistenciaTeste()
-    {
-        // Conexao aberta segurando o banco: fechada, o SQLite em memoria some.
-        _conexao = new SqliteConnection("DataSource=:memory:");
-        _conexao.Open();
-
-        _opcoes = new DbContextOptionsBuilder<CopilotoDbContext>()
-            .UseSqlite(_conexao)
-            .Options;
-
-        using var ctx = new CopilotoDbContext(_opcoes);
-        ctx.Database.EnsureCreated();
-    }
-
-    public void Dispose() => _conexao.Dispose();
-
-    private CopilotoDbContext Novo() => new(_opcoes);
+    private CopilotoDbContext Novo() => NovoContexto();
 
     [Fact]
     public void O_modelo_fecha_e_o_esquema_e_criado()
@@ -171,184 +150,5 @@ public class PersistenciaTeste : IDisposable
             .GetIndexes().Single(i => i.IsUnique);
 
         Assert.Equal("ux_leads_telefone", indice.GetDatabaseName());
-    }
-}
-
-/// <summary>
-/// O resolvedor sobre o banco (#103): a resolucao de Lead da #22 atravessando
-/// a persistencia, que e onde ela passa a valer entre reinicios.
-/// </summary>
-public class ResolvedorSobreBancoTeste : IDisposable
-{
-    private static readonly DateTimeOffset Agora = new(2026, 9, 1, 12, 0, 0, TimeSpan.Zero);
-    private const string Empresa = "+55 11 3333-4444";
-
-    private readonly SqliteConnection _conexao;
-    private readonly DbContextOptions<CopilotoDbContext> _opcoes;
-
-    public ResolvedorSobreBancoTeste()
-    {
-        _conexao = new SqliteConnection("DataSource=:memory:");
-        _conexao.Open();
-        _opcoes = new DbContextOptionsBuilder<CopilotoDbContext>().UseSqlite(_conexao).Options;
-        using var ctx = new CopilotoDbContext(_opcoes);
-        ctx.Database.EnsureCreated();
-    }
-
-    public void Dispose() => _conexao.Dispose();
-
-    [Fact]
-    public void O_lead_criado_hoje_e_encontrado_depois_do_restart()
-    {
-        // O que faltava para ser CRM: sem isto, o vendedor abre a tela e a
-        // conversa de ontem nao esta la.
-        Guid id;
-        using (var ctx = new CopilotoDbContext(_opcoes))
-        {
-            var r = new ResolvedorDeLead(Empresa, new LeadsNoBanco(ctx));
-            id = r.Resolver(Telefone.Normalizar("11 98765-4321")!, Agora).Id;
-        }
-
-        // Contexto novo = processo novo, para o efeito deste teste.
-        using (var ctx = new CopilotoDbContext(_opcoes))
-        {
-            var r = new ResolvedorDeLead(Empresa, new LeadsNoBanco(ctx));
-            var denovo = r.Resolver(Telefone.Normalizar("(11) 98765-4321")!, Agora);
-
-            Assert.Equal(id, denovo.Id);
-            Assert.Equal(1, r.LeadsConhecidos);
-        }
-    }
-
-    [Fact]
-    public void O_numero_sem_o_nono_digito_acha_o_lead_que_ja_esta_no_banco()
-    {
-        // A #22 atravessando o banco: normalizar no codigo so serve se a busca
-        // tambem for pelo normalizado.
-        using var ctx = new CopilotoDbContext(_opcoes);
-        var r = new ResolvedorDeLead(Empresa, new LeadsNoBanco(ctx));
-
-        var a = r.Resolver(Telefone.Normalizar("11 98765-4321")!, Agora);
-        var b = r.Resolver(Telefone.Normalizar("11 8765-4321")!, Agora);
-
-        Assert.Equal(a.Id, b.Id);
-        Assert.Equal(1, ctx.Leads.Count());
-    }
-}
-
-/// <summary>A Ficha do Cliente atravessando o banco (#86).</summary>
-public class FichaNoBancoTeste : IDisposable
-{
-    private static readonly DateTimeOffset T0 = new(2026, 9, 1, 12, 0, 0, TimeSpan.Zero);
-
-    private readonly SqliteConnection _conexao;
-    private readonly DbContextOptions<CopilotoDbContext> _opcoes;
-
-    public FichaNoBancoTeste()
-    {
-        _conexao = new SqliteConnection("DataSource=:memory:");
-        _conexao.Open();
-        _opcoes = new DbContextOptionsBuilder<CopilotoDbContext>().UseSqlite(_conexao).Options;
-        using var ctx = new CopilotoDbContext(_opcoes);
-        ctx.Database.EnsureCreated();
-    }
-
-    public void Dispose() => _conexao.Dispose();
-
-    [Fact]
-    public void A_ficha_volta_com_os_campos_preenchidos()
-    {
-        var id = Guid.NewGuid();
-        using (var ctx = new CopilotoDbContext(_opcoes))
-        {
-            var ficha = new Fichas.FichaCliente(id, Guid.NewGuid(), T0);
-            ficha.Atualizar(T0,
-                empresa: new Fichas.SobreAEmpresa(Ramo: Fichas.Anotacao.Fato("cafeteria"), Porte: Fichas.Anotacao.Fato("3 lojas")),
-                pessoa: new Fichas.SobreAPessoa(Cargo: Fichas.Anotacao.Fato("sócio")));
-            ctx.Fichas.Add(ficha);
-            ctx.SaveChanges();
-        }
-
-        using var leitura = new CopilotoDbContext(_opcoes);
-        var lida = leitura.Fichas.Single(f => f.Id == id);
-
-        Assert.Equal("cafeteria", lida.Empresa.Ramo!.Valor);
-        Assert.Equal("3 lojas", lida.Empresa.Porte!.Valor);
-        Assert.Equal("sócio", lida.Pessoa.Cargo!.Valor);
-        Assert.False(lida.EstaVazia);
-    }
-
-    [Fact]
-    public void O_historico_sobrevive_ao_banco()
-    {
-        // "Ele era o decisor e agora nao e" so vale se durar mais que a sessao.
-        var id = Guid.NewGuid();
-        using (var ctx = new CopilotoDbContext(_opcoes))
-        {
-            var ficha = new Fichas.FichaCliente(id, Guid.NewGuid(), T0);
-            ficha.Atualizar(T0, pessoa: new Fichas.SobreAPessoa(PapelNaDecisao: Fichas.Anotacao.Fato("decisor")));
-            ficha.Atualizar(T0.AddDays(2), pessoa: new Fichas.SobreAPessoa(PapelNaDecisao: Fichas.Anotacao.Fato("influenciador")));
-            ctx.Fichas.Add(ficha);
-            ctx.SaveChanges();
-        }
-
-        using var leitura = new CopilotoDbContext(_opcoes);
-        var lida = leitura.Fichas.Single(f => f.Id == id);
-
-        Assert.Equal(2, lida.Historico.Count);
-        Assert.Equal("decisor", lida.Historico[0].Pessoa.PapelNaDecisao!.Valor);
-    }
-
-    [Fact]
-    public void Ficha_vazia_e_gravavel()
-    {
-        // O sistema funciona sem ela, e "funciona" inclui salvar.
-        using var ctx = new CopilotoDbContext(_opcoes);
-        ctx.Fichas.Add(new Fichas.FichaCliente(Guid.NewGuid(), Guid.NewGuid(), T0));
-
-        ctx.SaveChanges();
-
-        Assert.True(ctx.Fichas.Single().EstaVazia);
-    }
-
-    [Fact]
-    public void A_natureza_da_anotacao_sobrevive_ao_banco()
-    {
-        // O conversor JSON e' quem reconstroi a Anotacao na leitura, e se ele
-        // errasse a natureza a impressao voltaria do banco como FATO — sem
-        // erro, sem log, e ancorando preco na proxima analise.
-        var id = Guid.NewGuid();
-        using (var ctx = new CopilotoDbContext(_opcoes))
-        {
-            var ficha = new Fichas.FichaCliente(id, Guid.NewGuid(), T0);
-            ficha.Atualizar(T0, pessoa: new Fichas.SobreAPessoa(
-                Cargo: Fichas.Anotacao.Fato("sócio", "LinkedIn"),
-                EstiloObservado: Fichas.Anotacao.Impressao("parece desconfiado", T0)));
-            ctx.Fichas.Add(ficha);
-            ctx.SaveChanges();
-        }
-
-        using var leitura = new CopilotoDbContext(_opcoes);
-        var lida = leitura.Fichas.Single(f => f.Id == id);
-
-        Assert.True(lida.Pessoa.Cargo!.EhFato);
-        Assert.Equal("LinkedIn", lida.Pessoa.Cargo!.Fonte);
-        Assert.False(lida.Pessoa.EstiloObservado!.EhFato);
-        Assert.Equal(T0, lida.Pessoa.EstiloObservado!.Quando);
-        Assert.Single(lida.Impressoes);
-    }
-
-    [Fact]
-    public void Um_lead_nao_tem_duas_fichas()
-    {
-        // Duas seriam duas versoes da verdade sem criterio de desempate.
-        var lead = Guid.NewGuid();
-        using var ctx = new CopilotoDbContext(_opcoes);
-        ctx.Fichas.Add(new Fichas.FichaCliente(Guid.NewGuid(), lead, T0));
-        ctx.SaveChanges();
-
-        ctx.Fichas.Add(new Fichas.FichaCliente(Guid.NewGuid(), lead, T0));
-
-        Assert.Throws<DbUpdateException>(() => ctx.SaveChanges());
     }
 }
